requirements.txt
asgiref==3.3.1
Django==3.1.3
django-ckeditor==6.0.0
django-crispy-forms==1.11.0
django-js-asset==1.2.2
packaging==20.9
Pillow==8.1.0
pyparsing==2.4.7
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
webencodings==0.5.1
